+++ src/opentelemetry/launcher/configuration.py
@@ -24,12 +24,16 @@ from logging import (
     getLevelName,
     getLogger,
 )
+from socket import gethostname
 from typing import Optional
 
 from environs import Env
 from grpc import ssl_channel_credentials
 
 from opentelemetry.baggage.propagation import BaggagePropagator
+from opentelemetry.trace.propagation.tracecontext import (
+    TraceContextTextMapPropagator,
+)
 from opentelemetry.instrumentation.instrumentor import BaseInstrumentor
 from opentelemetry.launcher.tracer import LightstepOTLPSpanExporter
 from opentelemetry.launcher.version import __version__
@@ -76,6 +80,11 @@ _OTEL_EXPORTER_OTLP_METRIC_INSECURE = _env.bool(
     "OTEL_EXPORTER_OTLP_METRIC_INSECURE", False
 )
 
+# FIXME Find a way to "import" this value from:
+# https://github.com/open-telemetry/opentelemetry-collector/blob/master/translator/conventions/opentelemetry.go
+# instead of hardcoding it here.
+_ATTRIBUTE_HOST_NAME = "host.name"
+
 
 class InvalidConfigurationError(Exception):
     """
@@ -97,6 +106,7 @@ def configure_opentelemetry(
     _auto_instrumented: bool = False,
 ):
     # pylint: disable=too-many-locals
+    # pylint: disable=too-many-statements
     """
     Configures OpenTelemetry with Lightstep environment variables
 
@@ -250,7 +260,11 @@ def configure_opentelemetry(
     set_global_textmap(
         CompositeHTTPPropagator(
             [
-                {"b3": B3Format(), "baggage": BaggagePropagator()}[propagator]
+                {
+                    "b3": B3Format(),
+                    "baggage": BaggagePropagator(),
+                    "tracecontext": TraceContextTextMapPropagator(),
+                }[propagator]
                 for propagator in propagators
             ]
         )
@@ -288,6 +302,27 @@ def configure_opentelemetry(
         )
     )
 
+    if _ATTRIBUTE_HOST_NAME not in resource_attributes.keys() or not (
+        resource_attributes[_ATTRIBUTE_HOST_NAME]
+    ):
+
+        no_hostname_message = (
+            "set it with the environment variable OTEL_RESOURCE_ATTRIBUTES or "
+            'with the resource_attributes argument. Use "host.name" as key '
+            "in both cases."
+        )
+        try:
+            hostname = gethostname()
+            if not hostname:
+                _logger.warning("Hostname is empty, %s", no_hostname_message)
+            else:
+                resource_attributes[_ATTRIBUTE_HOST_NAME] = hostname
+        # pylint: disable=broad-except
+        except Exception:
+            _logger.exception(
+                "Unable to get hostname, %s", no_hostname_message
+            )
+
     get_tracer_provider().resource = Resource(resource_attributes)
 
     if log_level >= DEBUG:
@@ -310,7 +345,11 @@ class LightstepLauncherInstrumentor(BaseInstrumentor):
             configure_opentelemetry(_auto_instrumented=True)
         except InvalidConfigurationError:
             _logger.exception(
-                "application instrumented via opentelemetry-instrument. all required configuration must be set via environment variables"
+                (
+                    "application instrumented via opentelemetry-instrument. "
+                    "all required configuration must be set via environment "
+                    "variables"
+                )
             )
 
     def _uninstrument(self, **kwargs):