+++ tests/test_configuration.py
@@ -22,6 +22,7 @@ from opentelemetry.launcher.configuration import (
     configure_opentelemetry,
     _logger,
     InvalidConfigurationError,
+    _ATTRIBUTE_HOST_NAME,
 )
 from opentelemetry.launcher.version import __version__
 from opentelemetry import baggage, trace
@@ -138,8 +139,11 @@ class TestConfiguration(TestCase):
                     log_level="WaRNiNG",
                 )
 
+    @patch("opentelemetry.launcher.configuration.gethostname")
     @patch("opentelemetry.launcher.configuration.Resource")
-    def test_resource_attributes(self, mock_resource):
+    def test_resource_attributes(self, mock_resource, mock_gethostname):
+        mock_gethostname.return_value = "the_hostname"
+
         configure_opentelemetry(
             service_name="service_name",
             service_version="service_version",
@@ -152,11 +156,15 @@ class TestConfiguration(TestCase):
                 "telemetry.sdk.version": __version__,
                 "service.name": "service_name",
                 "service.version": "service_version",
+                _ATTRIBUTE_HOST_NAME: "the_hostname",
             }
         )
 
+    @patch("opentelemetry.launcher.configuration.gethostname")
     @patch("opentelemetry.launcher.configuration.Resource")
-    def test_service_version(self, mock_resource):
+    def test_service_version(self, mock_resource, mock_gethostname):
+        mock_gethostname.return_value = "the_hostname"
+
         configure_opentelemetry(
             service_name="service_name",
             access_token="a" * 104,
@@ -167,6 +175,7 @@ class TestConfiguration(TestCase):
                 "telemetry.sdk.language": "python",
                 "telemetry.sdk.version": __version__,
                 "service.name": "service_name",
+                _ATTRIBUTE_HOST_NAME: "the_hostname",
             }
         )
 
@@ -196,7 +205,7 @@ class TestConfiguration(TestCase):
             propagators="baggage",
         )
 
-        with trace.get_tracer(__name__).start_as_current_span("test") as span:
+        with trace.get_tracer(__name__).start_as_current_span("test"):
             ctx = baggage.set_baggage("abc", "def")
             prop = get_global_textmap()
             carrier = {}
@@ -204,12 +213,32 @@ class TestConfiguration(TestCase):
             self.assertIsNone(carrier.get("x-b3-traceid"))
             self.assertEqual(carrier.get("baggage"), "abc=def")
 
+    def test_propagation_tracecontext(self):
+        configure_opentelemetry(
+            service_name="service_name",
+            service_version="service_version",
+            access_token="a" * 104,
+            propagators="tracecontext",
+        )
+
+        with trace.get_tracer(__name__).start_as_current_span("test") as span:
+            ctx = baggage.set_baggage("abc", "def")
+            prop = get_global_textmap()
+            carrier = {}
+            prop.inject(dict.__setitem__, carrier, context=ctx)
+            self.assertIn(
+                "00-{}".format(
+                    format(span.get_span_context().trace_id, "032x")
+                ),
+                carrier.get("traceparent"),
+            )
+
     def test_propagation_multiple(self):
         configure_opentelemetry(
             service_name="service_name",
             service_version="service_version",
             access_token="a" * 104,
-            propagators="b3,baggage",
+            propagators="b3,baggage,tracecontext",
         )
 
         with trace.get_tracer(__name__).start_as_current_span("test") as span:
@@ -222,3 +251,44 @@ class TestConfiguration(TestCase):
                 carrier.get("x-b3-traceid"),
             )
             self.assertEqual(carrier.get("baggage"), "abc=def")
+            self.assertIn(
+                "00-{}".format(
+                    format(span.get_span_context().trace_id, "032x")
+                ),
+                carrier.get("traceparent"),
+            )
+
+    @patch("opentelemetry.launcher.configuration.gethostname")
+    @patch("opentelemetry.launcher.configuration.Resource")
+    def test_hostname(self, mock_resource, mock_gethostname):
+
+        mock_gethostname.return_value = "the_hostname"
+
+        configure_opentelemetry(
+            service_name="service_name",
+            access_token="a" * 104,
+        )
+
+        mock_resource.assert_called_with(
+            {
+                "telemetry.sdk.language": "python",
+                "telemetry.sdk.version": __version__,
+                "service.name": "service_name",
+                _ATTRIBUTE_HOST_NAME: "the_hostname",
+            }
+        )
+
+        configure_opentelemetry(
+            service_name="service_name",
+            access_token="a" * 104,
+            resource_attributes="{}=other_hostname".format(
+                _ATTRIBUTE_HOST_NAME
+            )
+        )
+
+        mock_resource.assert_called_with(
+            {
+                "service.name": "service_name",
+                _ATTRIBUTE_HOST_NAME: "other_hostname",
+            }
+        )